Don't push down functions by default
Make builtin function pushdown opt-in: Postgres builtins now ship to
ClickHouse only when explicitly mapped, so name or signature differences
cannot silently alter results.

Add explicit mappings for `mod`, `pow`/`power`, `bit_count(bytea)`, and
`reverse(text)` (→ `reverseUTF8`) to retain previously working
pushdowns.

Fix `editDistance`, `length(text)` and `strpos(text, text)` pushdown to
map to `lengthUTF8` and `positionUTF8` rather than ClickHouse's
byte-counting `length` and `position`, matching Postgres character
semantics.

Stop pushing down `asin`, `acos`, `atanh`, and `acosh`: Postgres raises
an error on out-of-range input where ClickHouse returns `NaN`.
Evaluating locally preserves Postgres semantics.

Add comprehensive tests for all of these behaviors.

+++ src/binary.cpp
@@ -873,7 +873,7 @@ extern "C"
 
 		/* Initialize coltypes to SELECT types, when provided. */
 		if (query->tupdesc)
-			for (size_t i = 0; i < list_length(query->attr_nums); i++)
+			for (size_t i = 0; i < (size_t)list_length(query->attr_nums); i++)
 				state->coltypes[i]
 				= TupleDescAttr(query->tupdesc, lfirst_int(list_nth_cell(query->attr_nums, i)) - 1)->atttypid;
 		else
